backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
PyJWT==2.10.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
cryptography==41.0.7
djangorestframework-simplejwt==5.5.1
django-filter==23.1.0
whitenoise==6.6.0
gunicorn==21.2.0

